Circle fixes: move async layout state init
