+++ src/ui/setup/index.ts
@@ -13,10 +13,17 @@ import { getUserInfo } from "ui/hooks/users";
 import { getUserSettings } from "ui/hooks/settings";
 import { initLaunchDarkly } from "ui/utils/launchdarkly";
 import { maybeSetMixpanelContext } from "ui/utils/mixpanel";
-import { getInitialLayoutState } from "ui/reducers/layout";
+import { syncInitialLayoutState } from "ui/reducers/layout";
 import { getInitialTabsState } from "devtools/client/debugger/src/reducers/tabs";
 import { getInitialCommentsState } from "ui/reducers/comments";
 import { initialMessageState } from "devtools/client/webconsole/reducers/messages";
+import { trackEvent } from "ui/utils/telemetry";
+import { Recording } from "ui/types";
+import { getRecording } from "ui/hooks/recordings";
+import { getRecordingId } from "ui/utils/recording";
+import { getReplaySession } from "ui/setup/prefs";
+import type { LayoutState } from "ui/state/layout";
+
 declare global {
   interface Window {
     store: UIStore;
@@ -26,6 +33,62 @@ declare global {
 // Just to grab the type of Dispatch
 let store: UIStore;
 export type AppDispatch = typeof store.dispatch;
+
+const getDefaultSelectedPrimaryPanel = (session: any, recording?: Recording) => {
+  if (session) {
+    return session.selectedPrimaryPanel;
+  }
+
+  if (!recording) {
+    return syncInitialLayoutState.selectedPrimaryPanel;
+  }
+
+  return recording.comments.length ? "comments" : syncInitialLayoutState.selectedPrimaryPanel;
+};
+
+export async function getInitialLayoutState(): Promise<LayoutState> {
+  const recordingId = getRecordingId();
+
+  // If we're in the library, there are no preferences to fetch.
+  if (!recordingId) {
+    return syncInitialLayoutState;
+  }
+
+  let recording;
+  try {
+    recording = await getRecording(recordingId);
+  } catch (e) {
+    return syncInitialLayoutState;
+  }
+
+  const session = await getReplaySession(recordingId);
+
+  if (!session) {
+    return {
+      ...syncInitialLayoutState,
+      selectedPrimaryPanel: getDefaultSelectedPrimaryPanel(session, recording),
+    };
+  }
+
+  const { viewMode, showVideoPanel, toolboxLayout, selectedPanel, consoleFilterDrawerExpanded } =
+    syncInitialLayoutState;
+  const initialViewMode = session.viewMode || viewMode;
+  trackEvent(initialViewMode == "dev" ? "layout.default_devtools" : "layout.default_viewer");
+
+  return {
+    ...syncInitialLayoutState,
+    consoleFilterDrawerExpanded:
+      "consoleFilterDrawerExpanded" in session
+        ? session.consoleFilterDrawerExpanded
+        : consoleFilterDrawerExpanded,
+    viewMode: initialViewMode,
+    selectedPanel: "selectedPanel" in session ? session.selectedPanel : selectedPanel,
+    selectedPrimaryPanel: getDefaultSelectedPrimaryPanel(session, recording),
+    showVideoPanel: "showVideoPanel" in session ? session.showVideoPanel : showVideoPanel,
+    toolboxLayout: "toolboxLayout" in session ? session.toolboxLayout : toolboxLayout,
+    localNags: "localNags" in session ? session.localNags : [],
+  };
+}
 
 export async function bootstrapApp() {
   const initialState = {
